extract BulkActions
